NUTCH-1714 Nutch 2.x upgrade to Gora 0.4
git-svn-id: https://svn.apache.org/repos/asf/nutch/branches/2.x@1594812 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/nutch/crawl/AbstractFetchSchedule.java
@@ -17,14 +17,14 @@
 
 package org.apache.nutch.crawl;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.nutch.storage.WebPage;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * This class provides common methods for implementations of
@@ -196,7 +196,7 @@ implements FetchSchedule {
     // reduce fetchInterval so that it fits within the max value
     if (page.getFetchInterval() > maxInterval)
       page.setFetchInterval(Math.round(maxInterval * 0.9f));
-    page.setStatus(CrawlStatus.STATUS_UNFETCHED);
+    page.setStatus((int) CrawlStatus.STATUS_UNFETCHED);
     page.setRetriesSinceFetch(0);
     // TODO: row.setSignature(null) ??
     page.setModifiedTime(0L);
